# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: CRM ИТ Школы РТК — REST API
  version: 0.2.0
  description: |
    API веб-клиента CRM и точка интеграции с корпоративными системами.

    Аутентификация: вход через Keycloak (GET /api/auth/login) создаёт серверную сессию BFF, браузер получает HttpOnly cookie SESSION; токены OIDC клиенту не передаются. Выход (POST /api/auth/logout) завершает сессию CRM и возвращает адрес завершения сессии Keycloak.

    Изменяющие запросы передают заголовок X-CSRF-TOKEN со значением token из GET /api/csrf. Команды передают Idempotency-Key; повтор с тем же ключом и телом возвращает прежний результат, с другим телом — 409 IDEMPOTENCY_CONFLICT. Устаревшая version даёт 409 VERSION_CONFLICT с currentVersion.

    Ошибки возвращаются в формате ApiError {code, message, requestId}. requestId совпадает с заголовком X-Request-Id и записью в журнале сервера; stack trace и секреты в ответ не попадают.
servers:
  - url: /
...
paths:
...
  /api/interactions:
    get:
      tags: [Interaction]
      summary: Взаимодействия видимых вузов
      operationId: listInteractions
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/InteractionOrganizationId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/InteractionSort'
        - name: q
          in: query
          required: false
          description: Часть названия взаимодействия или вуза без учёта регистра
          schema:
            type: string
            maxLength: 200
        - name: due
          in: query
          required: false
          description: OVERDUE — шаг и срок заданы, срок прошёл; THIS_WEEK — шаг и срок заданы, срок на текущей неделе по Москве (понедельник–воскресенье); NO_NEXT_STEP — нет текста шага или его срока, такие взаимодействия попадают только сюда
          schema:
            type: string
            enum: [OVERDUE, THIS_WEEK, NO_NEXT_STEP]
        - name: stage
          in: query
          required: false
          description: Точное название текущего этапа
          schema:
            type: string
            maxLength: 200
      responses:
        '200':
          description: Interactions of all visible organizations, or of one visible organization when organizationId is given; filters, total and pages are computed on the server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageInteraction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Interaction]
      summary: Создать взаимодействие
      operationId: createInteraction
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCreate'
      responses:
        '201':
          description: Interaction created with a private 13-stage workflow snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ContactInteractionMutationForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/interactions/{id}:
    get:
      tags: [Interaction]
      summary: Карточка взаимодействия
      operationId: getInteraction
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Interaction visible through its current organization scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Interaction]
      summary: Изменить план взаимодействия
      operationId: updateInteractionPlan
      x-implementation-stage: 3
      description: >-
        Changes the plan of the card: next step, its due date, program and products.
        An absent property keeps the current value, an explicit null clears it.
        Removing a product whose agreement already has contract data is rejected.
        A change writes one PLAN_UPDATED history event; a request without effective changes returns the card unchanged.
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionPlanUpdate'
      responses:
        '200':
          description: Interaction after the plan update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ContactInteractionMutationForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/interactions/{id}/events:
    get:
      tags: [Interaction]
      summary: История взаимодействия
      operationId: listInteractionEvents
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Immutable interaction history ordered by version, occurredAt and id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InteractionEvent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrmProfileRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/interactions/{id}/transitions:
    post:
      tags: [Interaction]
      summary: Перевести взаимодействие на этап
      operationId: transitionInteraction
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionTransition'
      responses:
        '200':
          description: Interaction after a validated workflow transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ContactInteractionMutationForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/interactions/{id}/comments:
    post:
      tags: [Interaction]
      summary: Добавить комментарий к взаимодействию
      operationId: commentInteraction
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionComment'
      responses:
        '200':
          description: Updated interaction and the immutable comment event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionCommentResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ContactInteractionMutationForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/interactions/{id}/stage-edits:
    post:
      tags: [Interaction]
      summary: Изменить этапы взаимодействия
      operationId: editInteractionStages
      x-implementation-stage: 3
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionStageEdit'
      responses:
        '200':
          description: >-
            Interaction after a validated local workflow snapshot edit. Structural edits rebuild only the
            transitions around the added, moved or deleted stage and write one STAGES_EDITED history event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ContactInteractionMutationForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/interactions/{id}/attachments:
    post:
      tags: [Attachment]
      summary: Загрузить вложение
      operationId: uploadInteractionAttachment
      x-implementation-stage: 1
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, stageId]
              properties:
                file:
                  type: string
                  format: binary
                stageId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Attachment stored in quarantine and scanned before any possible download
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ContactInteractionMutationForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
      description: Серверная сессия BFF, созданная входом через GET /api/auth/login. Swagger UI на том же origin использует сессию браузера.
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    InteractionOrganizationId:
      name: organizationId
      in: query
      required: false
      schema:
        type: string
        format: uuid
    InteractionSort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: ['createdAt,asc', 'createdAt,desc', 'updatedAt,asc', 'updatedAt,desc', 'nextActionAt,asc']
        default: 'updatedAt,desc'
...
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Значение token из GET /api/csrf для текущей сессии. Без заголовка или с чужим значением запрос отклоняется с 403 FORBIDDEN.
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
  responses:
    BadRequest:
      description: Validation or request format error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: BFF session is absent or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    CrmProfileRequired:
      description: The OIDC identity has no active CRM profile (CRM_PROFILE_REQUIRED) or its profile created at first login waits for administrator activation (CRM_PROFILE_PENDING)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ContactInteractionMutationForbidden:
      description: A visible organization or interaction can be changed only by an active USER or LEADER of its scope; the scope is checked first, so a hidden object is 404. Also returned when the OIDC identity has no active CRM profile
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Object does not exist or is outside the current server-side scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Version or idempotency conflict; VERSION_CONFLICT includes currentVersion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    PayloadTooLarge:
      description: Uploaded payload exceeds the configured limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalError:
      description: Unexpected server error; the stack trace is written only to the server log under the same requestId
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHENTICATED
            - FORBIDDEN
            - CRM_PROFILE_REQUIRED
            - CRM_PROFILE_PENDING
            - LAST_ACTIVE_ADMIN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VERSION_CONFLICT
            - IDEMPOTENCY_CONFLICT
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - RATE_LIMITED
            - INTERNAL_ERROR
            - DEPENDENCY_UNAVAILABLE
            - REPORT_NOT_READY
            - REPORT_ACCESS_CHANGED
            - REPORT_RESULT_UNAVAILABLE
            - REPORT_CAPACITY_EXCEEDED
            - SOURCE_NOT_CONFIGURED
            - SYNC_ALREADY_RUNNING
            - SYNC_CAPACITY_EXCEEDED
            - LMS_NOT_MAPPED
            - LMS_SYNC_FAILED
        message:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        currentVersion:
          type: integer
          minimum: 0
...
    PageMeta:
      type: object
      required: [page, size, total]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
...
    CatalogReference:
      type: object
      required: [id, name, archived, version]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        archived:
          type: boolean
        version:
          type: integer
          minimum: 0
...
    ProductAgreement:
      type: object
      required: [id, productId, productName, productArchived, contractNumber, licenseSigned, licenseExpiryYear, transferStatus]
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productName:
          type: string
        productArchived:
          type: boolean
        contractNumber:
          type: [string, 'null']
        licenseSigned:
          type: [boolean, 'null']
        licenseExpiryYear:
          type: [integer, 'null']
        transferStatus:
          type: [string, 'null']
    InteractionStage:
      type: object
      required: [id, name, order, optional]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        order:
          type: integer
          minimum: 0
        optional:
          type: boolean
    InteractionTransitionOption:
      type: object
      required: [stageId, stageName, commentRequired]
      properties:
        stageId:
          type: string
          format: uuid
        stageName:
          type: string
        commentRequired:
          type: boolean
    InteractionStageTransition:
      type: object
      required: [fromStageId, toStageId, commentRequired]
      properties:
        fromStageId:
          type: string
          format: uuid
        toStageId:
          type: string
          format: uuid
        commentRequired:
          type: boolean
    InteractionSummary:
      type: object
      required: [id, organizationId, title, currentStageId, currentStageName, nextAction, nextActionAt, contactIds, programId, productIds, lastContactAt, version, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        title:
          type: string
        currentStageId:
          type: string
          format: uuid
        currentStageName:
          type: string
        nextAction:
          type: [string, 'null']
        nextActionAt:
          type: [string, 'null']
          format: date-time
        contactIds:
          type: array
          items:
            type: string
            format: uuid
        programId:
          type: [string, 'null']
          format: uuid
        productIds:
          type: array
          items:
            type: string
            format: uuid
        lastContactAt:
          type: [string, 'null']
          format: date-time
        version:
          type: integer
          minimum: 0
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Interaction:
      allOf:
        - $ref: '#/components/schemas/InteractionSummary'
        - type: object
          required: [stages, transitions, allowedTransitions, program, productAgreements, attachments]
          properties:
            stages:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/InteractionStage'
            transitions:
              type: array
              items:
                $ref: '#/components/schemas/InteractionStageTransition'
            allowedTransitions:
              type: array
              items:
                $ref: '#/components/schemas/InteractionTransitionOption'
            program:
              anyOf:
                - $ref: '#/components/schemas/CatalogReference'
                - type: 'null'
            productAgreements:
              type: array
              items:
                $ref: '#/components/schemas/ProductAgreement'
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
    PageInteraction:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/InteractionListItem'
    InteractionListItem:
      allOf:
        - $ref: '#/components/schemas/InteractionSummary'
        - type: object
          required: [organizationName, programName, ownerManagerName]
          properties:
            organizationName:
              type: string
            programName:
              type: [string, 'null']
            ownerManagerName:
              type: [string, 'null']
    InteractionCreate:
      type: object
      required: [organizationId, title]
      properties:
        organizationId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        nextAction:
          type: [string, 'null']
          maxLength: 500
        nextActionAt:
          type: [string, 'null']
          format: date-time
        contactIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
        programId:
          type: [string, 'null']
          format: uuid
        productIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
        lastContactAt:
          type: [string, 'null']
          format: date-time
        templateId:
          type: [string, 'null']
          format: uuid
    InteractionTransition:
      type: object
      required: [version, toStageId]
      properties:
        version:
          type: integer
          minimum: 0
        toStageId:
          type: string
          format: uuid
        comment:
          type: [string, 'null']
          maxLength: 4000
        attachmentIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
        nextStep:
          description: When present, replaces both the next step and its due date in the same command
          anyOf:
            - $ref: '#/components/schemas/InteractionNextStep'
            - type: 'null'
    InteractionComment:
      type: object
      required: [version, stageId, text]
      properties:
        version:
          type: integer
          minimum: 0
        stageId:
          type: string
          format: uuid
        text:
          type: string
          minLength: 1
          maxLength: 4000
        attachmentIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
        nextStep:
          description: When present, replaces both the next step and its due date in the same command
          anyOf:
            - $ref: '#/components/schemas/InteractionNextStep'
            - type: 'null'
    InteractionNextStep:
      type: object
      required: [nextAction, nextActionAt]
      properties:
        nextAction:
          type: [string, 'null']
          maxLength: 500
        nextActionAt:
          type: [string, 'null']
          format: date-time
    InteractionPlanUpdate:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          minimum: 0
        nextAction:
          type: [string, 'null']
          maxLength: 500
        nextActionAt:
          type: [string, 'null']
          format: date-time
        programId:
          type: [string, 'null']
          format: uuid
        productIds:
          type: [array, 'null']
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
    InteractionStageEdit:
      type: object
      required: [version, operations]
      properties:
        version:
          type: integer
          minimum: 0
        operations:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/InteractionStageEditOperation'
...
    Attachment:
      type: object
      required: [id, interactionId, stageId, eventId, originalName, mediaType, sizeBytes, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        interactionId:
          type: string
          format: uuid
        stageId:
          type: string
          format: uuid
        eventId:
          type: [string, 'null']
          format: uuid
        originalName:
          type: string
          maxLength: 255
        mediaType:
          type: string
        sizeBytes:
          type: integer
          format: int64
          minimum: 0
        status:
          type: string
          enum: [QUARANTINE, CLEAN, REJECTED, UNVERIFIABLE]
        createdAt:
          type: string
          format: date-time
    InteractionCommentResult:
      type: object
      required: [interaction, event]
      properties:
        interaction:
          $ref: '#/components/schemas/Interaction'
        event:
          $ref: '#/components/schemas/InteractionEvent'
    InteractionEvent:
      type: object
      required: [id, type, stageId, stageNameSnapshot, nextStep, actorProfileId, actorDisplayName, version, commandId, occurredAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [CREATED, TRANSITIONED, COMMENTED, STAGES_EDITED, PLAN_UPDATED]
        stageId:
          type: string
          format: uuid
        stageNameSnapshot:
          type: string
        fromStageId:
          type: [string, 'null']
          format: uuid
        fromStageNameSnapshot:
          type: [string, 'null']
        toStageId:
          type: [string, 'null']
          format: uuid
        toStageNameSnapshot:
          type: [string, 'null']
        comment:
          type: [string, 'null']
        nextStep:
          description: Next step and due date set by this event; null when the event did not change them
          anyOf:
            - $ref: '#/components/schemas/InteractionNextStep'
            - type: 'null'
        actorProfileId:
          type: string
          format: uuid
        actorDisplayName:
          type: string
        ownerManagerIdSnapshot:
          type: [string, 'null']
          format: uuid
        version:
          type: integer
          minimum: 0
        commandId:
          type: string
          format: uuid
        occurredAt:
          type: string
          format: date-time
...
security:
  - cookieAuth: []